Refactoring new loaders framework to Wildfly Core Loaders module.
