+++ core-feature-pack/src/main/resources/modules/system/layers/base/org/jboss/as/server/main/module.xml
@@ -92,5 +92,6 @@
         <module name="org.picketbox" optional="true"/>
         <module name="io.undertow.core" />
         <module name="org.wildfly.common"/>
+        <module name="org.wildfly.loaders"/>
     </dependencies>
 </module>
