+++ core-feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/loaders/main/module.xml
@@ -2,7 +2,7 @@
 
 <!--
   ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2010, Red Hat, Inc., and individual contributors
+  ~ Copyright 2015, Red Hat, Inc., and individual contributors
   ~ as indicated by the @author tags. See the copyright.txt file in the
   ~ distribution for a full listing of individual contributors.
   ~
@@ -22,27 +22,17 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<module xmlns="urn:jboss:module:1.3" name="org.jboss.as.logging">
+<module xmlns="urn:jboss:module:1.3" name="org.wildfly.loaders">
     <properties>
         <property name="jboss.api" value="private"/>
+        <property name="jboss.require-java-version" value="1.8"/>
     </properties>
 
     <resources>
-        <artifact name="${org.wildfly.core:wildfly-logging}"/>
+        <artifact name="${org.wildfly.core:wildfly-loaders}"/>
     </resources>
 
     <dependencies>
-        <module name="javax.api"/>
-        <module name="org.apache.log4j"/>
-        <module name="org.jboss.as.controller"/>
-        <module name="org.wildfly.security.elytron"/>
-        <module name="org.wildfly.loaders"/>
-        <module name="org.jboss.as.server"/>
-        <module name="org.jboss.staxmapper"/>
-        <module name="org.jboss.msc"/>
-        <module name="org.jboss.logging"/>
-        <module name="org.jboss.logmanager"/>
         <module name="org.jboss.modules"/>
-        <module name="org.jboss.stdio"/>
     </dependencies>
 </module>
