+++ loaders/src/main/java/org/wildfly/loaders/IOUtils.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.server.loaders;
+package org.wildfly.loaders;
 
 import org.jboss.modules.Resource;
