+++ loaders/src/test/java/org/wildfly/loaders/ResourceLoadersTest.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.server.loaders;
+package org.wildfly.loaders;
 
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertFalse;
