+++ loaders/src/test/resources/loader-test-exploded/foo.ear/META-INF/MANIFEST.MF
