+++ loaders/src/test/resources/loader-test-exploded/foo.ear/app.war/META-INF/MANIFEST.MF
