+++ loaders/src/test/resources/loader-test-exploded/foo.ear/app.war/WEB-INF/classes/my/company/GreetingServlet.class
