+++ loaders/src/test/resources/loader-test-exploded/foo.ear/app.war/WEB-INF/lib/lib.jar/generator/Messages.class
