+++ loaders/src/test/resources/loader-test-exploded/foo.ear/app.war/WEB-INF/web.xml
