+++ loaders/src/test/resources/loader-test-mixed1/foo.ear
