+++ loaders/src/test/resources/loader-test-mixed2/foo.ear/META-INF/MANIFEST.MF
