+++ loaders/src/test/resources/loader-test-mixed2/foo.ear/app.war
