+++ loaders/src/test/resources/loader-test-packaged/foo.ear
