+++ server/pom.xml
@@ -39,6 +39,10 @@
     <dependencies>
         <dependency>
             <groupId>org.wildfly.core</groupId>
+            <artifactId>wildfly-loaders</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wildfly.core</groupId>
             <artifactId>wildfly-controller</artifactId>
         </dependency>
         <dependency>
