+++ server/src/main/java/org/jboss/as/server/deployment/module/ResourceRoot.java
@@ -27,7 +27,7 @@ import java.util.List;
 
 import org.jboss.as.server.logging.ServerLogger;
 import org.jboss.as.server.deployment.SimpleAttachable;
-import org.jboss.as.server.loaders.ResourceLoader;
+import org.wildfly.loaders.ResourceLoader;
 
 /**
  * @author John E. Bailey
